+++ doc/changes.rst
@@ -2,12 +2,20 @@
 Change Log
 ==========
 
-3.5.1 (not released yet)
+3.6.0 (not released yet)
 ------------------------
 
+* Dynamically split ``FASTSPEC`` HDU into an additional ``MORELINES``
+  HDU [`PR #273`_].
+* Switch to new default templates (version 3.0.0) built from a
+  higher-resolution (``R=10000``) C3K library [`PR #271`_].
+* Pass ``constraintsfile`` argument to ``fastqa`` [`PR #270`_].
 * New unit test and documentation of external photometric catalog
   "mode" [`PR #269`_].
 
+.. _`PR #273`: https://github.com/desihub/fastspecfit/pull/273
+.. _`PR #271`: https://github.com/desihub/fastspecfit/pull/271
+.. _`PR #270`: https://github.com/desihub/fastspecfit/pull/270
 .. _`PR #269`: https://github.com/desihub/fastspecfit/pull/269
 
 3.5.0 (2026-06-16)